+++ src/@types/component.d.ts
@@ -1,0 +1,3 @@
+interface RedioFromReference {
+  checked(target: Number) : void
+}